Refine "other versions" behavior (#1538)
When you click the "other versions" option it used to take you to the
version link farm page. While it is nice to get all the versions this
loses the one of the nice behaviors of the versions drop down, namely
that it checks if the page you are on exists in the target version and
drops you right on that page if it does.

This change stops it from sending you to the link farm and instead
un-hides another drop down that contains *all* of the versions.

Relates to #1419 and #1536.