Improve rendering of callout lists in direct_html (#1582)
When you write a callout list with duplicate entries we'd previously
duplicate a bunch of the callouts in strange ways. Both docbook and
direct_html had different, strange renderings. This changes direct_html
to have a fairly sane rendering. Say you write this, like we do in the
Elasticsearch reference:
```
      "tokens" : [ {
        "token" : "detail",
        "start_offset" : 0,
        "end_offset" : 8,
        "type" : "<ALPHANUM>",
        "position" : 0,
        "keyword" : false <1>
      }, {
        "token" : "output",
        "start_offset" : 9,
        "end_offset" : 15,
        "type" : "<ALPHANUM>",
        "position" : 1,
        "keyword" : false <1>
      } ]
```

The duplicate `<1>` elemenets now get a single callout which looks sort
of like:
```
<1> Output only "keyword" attribute, since specify "attributes" in
the request.
```

This is *much* cleaner than how docbook does it and how direct_html used
to do it.