Make direct_html tables a little more docbook (#1591)
Apparently docbook only uses the formal "table" wrapper when there is a
title and uses the informal "informaltable" wrapper even when the table
has an id. direct_html was using the formal wrapper when it *either* had
an id or a title. This fixes that.

Required by the Elaticsearch reference (#1566).