Fix support for imagesdir
This fixes our image copier to properly copy images referenced by
`imagesdir`. We *were* copying the images correctly be we weren't
copying them to the right spot. This is because our image copier does a
`find | grep <path>` to find the image for compatibility with the old
`a2x` utility that we used to build the asciidoc before asciidoctor. We
just weren't getting the paths right.

Here is some documentation for the attribute:
https://asciidoctor.org/docs/user-manual/#setting-the-location-of-images