Fix locale errors in direct_html (#1649)
When attempting to switch some Chinese books we were getting errors that
looked like:
```
cli/options.rb:87:in `encode': "\xE8" from ASCII-8BIT to UTF-8
```

This was caused by ruby thinking that the cli was in ascii instead of
utf8 but perl encoding stuff in utf8. Ruby was doing that, helpfully,
because the shell's locale *was* `POSIX` because that is the default in
minideb if you don't run `locale-gen`.

So, interestingly, the fix for this is in the `Dockerfile`. We have to
enable the `en_US.UTF-8` locale. Now ruby reads the CLI as UTF-8.