Only build with `--direct_html` (#1659)
This removes support for building docs without direct_html. You no
longer need to specify `--direct_html` on the command line or
`direct_html` in conf.yaml for it ot build with direct_html.

This doesn't drop the configuration from conf.yaml or the change
tracker or all of the files that we need to build with docbook. We'll do
that in a followup.