Add rust client to Elasticsearch clients docs
This commit adds the Rust client to the Elasticsearch Clients documentation.

Per Rust crate idioms, the core documentation is hosted on docs.rs, which for the Rust client package is
at https://docs.rs/elasticsearch. The inclusion of the Rust client in the Clients documentation is to provide
representation in the official docs and an overview of the client.