Add Github link support to beta and experimental macros
We add the `experimental::[]` and `beta::[]` macros to document features
that are not yet ready for GA. However, these tags don't currently
indicate what steps are needed to get a feature to GA.

With these changes, users can include a GitHub issue for the feature in
the `experimental::[]` or `beta::[]`. When provided, a concise sentence
and link is generated for the issue.