[DOCS] Update attributes for multi arg footnotes
The `{multi-arg}` attribute is used in the function reference template
to add footnotes for parameters can be passed multiple times.

However, using this attribute multiple times on the same page results
in duplicated footnotes.

Changes:
* Updates the `{multi-arg}` attribute to use the `footnoteref`
syntax. This gives the footnote a reusable ID.

* Adds a new attribute, `{multi-arg-ref}`, which can be used for
the subsequent occurrences of the footnote.

* Updates guidance on usage of the `{multi-arg}` attribute
in the function reference template.