Use the "version" rather than "branch" in file paths
There's an existing method to give each "branch" a "title". The title
corresponds to the "name" of the version, in cases where it doesn't
match the name of the branch itself, and the "title" already appears in
the version dropdown. This change extends that to use the version name
("title") in the URL.

Future work will better distinguish between "branches" and "versions"
in cases where they aren't the same.

Related to https://github.com/elastic/docs/issues/1850