Simplify test contexts around building one book twice
Tests were failing as a result of the previous commit. After the list of
branches got updated, it was checking that the first build built all
the branches in the final list of branches (which it obviously shouldn't
do).

All of the tests in the "build all books" tests are already covered by
other examples, so we don't lose any coverage by removing them here.