Don't fail when trying to remove TOC from custom title page
In #1909, we added the ability to replace the title page of a book with
completely custom content. For multi-branch builds, we previously
assumed the index.html file would always contain the TOC, which caused
errors in Kibana builds when removing a non-existant TOC.

Also, we previously customized the title page and then tried to extract the
TOC. Now, we extract the TOC first, since it may not exist if the title
page has been customized.