added 6.x branch to elastic/apm-agent-python
We started releasing 6.x versions some time ago, but forgot about
updating the docs.

I left 5.x as a live branch as it is the last version of the agent
to support Python 2.7. 6.x+ only supports Python 3.

We added an item to our release checklist to avoid this from happening
going forward. See
https://github.com/elastic/apm-agent-python/pull/1109