Hide X-Pack tags on versions 7.11 and later
Eventually, we'll want to remove the x-pack role from the Asciidoc for
all sections/headers that include it, but until then this will at least
hide the visual indicator and link. The `xpack` class will be included
on the surrounding divs, since this is the default Asciidoctor behavior
for "role" tags.

Fixes #2207.