Fix Rubocop
For some of the metrics, I just bumped the maximum values slightly above
the values reported in the code. If I had more time, i would make some
of the methods simpler, but this isn't the time for significant
refactors.
I also changed the order in the Makefile to run style checks sooner,
before the integration tests run.