Use relative links on the landing page
This will keep the automatic link checker from checking these pages, but
is needed to get a fully-functional preview branch prior to this being
pushed live. Without this change, the link checker actually still works,
because it cuts off the domain prefix before looking for the existence
of files (so the URL doesn't work even though the file is there).