fix: '_find_toplevel' fails due to "fatal: detected dubious ownership in repository" error from latest git
A recent git update: add an owner check for the top-level directoy
  https://github.com/git/git/commit/8959555cee7ec045958f9b6dd62e541affb7e7d9
This was breaking `git rev-parse --show-toplevel` call made in
build_docs.pl's `sub _find_toplevel` with the message:
    INFO:build_docs:fatal: detected dubious ownership in repository at '<some dir>'
for some (macos?) users.

Removing the `--user $localUid:$localGid` for the `docker run ...` that
runs the doc build avoids this error by running the build as root
instead of as the local user (which is some UID that doesn't necessarily
map to a user in the docker container).

I am not sure there isn't some fallout from this change.