Tell git to treat all directories as safe
With a recent change to git [1], running commands like `git rev-parse
--show-toplevel` will fail when different directories are owned by
different users. Because of the complexity of how we use Docker and
expect it to work across different operating systems, the easiest
approach is to restore the prior behavior. [2]

Rather than trying to run `git config` from within the build process
(which runs into its own file permission errors depending on who is
running the command), instead we just hardcode the "global" gitconfig
file (which is located at the root of the Docker container).

[1]: https://github.com/git/git/commit/8959555cee7ec045958f9b6dd62e541affb7e7d9
[2]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-safedirectory