Create a dedicated build pipeline for PR
This commit  introduces a new build pipeline just for PR, without for
now renaming the underlying scripts (next commit).

We were using a single pipeline for commits to branches/master and for
PR. The issue is that all PR builds are handled by the Buildkite PR bot,
and this was introducing a lot of un-necessary complexity into the build
script to support both mode. Additionally, some builds were kicked-off
on commits to a branch of a PR interefering with the PR bot builds.