Use bigger CI machine for docs repo
We always do a full rebuild on `elastic/docs` which takes up to 2 hours
with the standard n1-standard-4 CI machine. This PR introduces an
optimization only for builds on elastic/docs to use n2-standard-32 to
drop the build time to about 1 hour. For any other builds, we'll use
n2-standard-4 and tolerate the <20 min build time for docs given the
price difference (10x cheaper).