Add base path to default URL
At some point in the 5 years since its original implementation, the Kibana team adopted a pattern for local development where a 3-letter  “base path” was added to the URL, I assume to help with caching issues and other things. This means that the link was broken for most users.

You can disable this by passing the flag `—no-base-path` when booting up. However, the console barks at you if you do this and strongly recommends against it.

For anyone running with that —no-base-path flaf, the console did work, but I would assume that is a small subset of users, as that flag is not documented, AFAICT.
I’ve asked if there is a way I can have a magic link that works for both, but I don’t believe there will be.

This change accounts for the newer way of doing it, (with the base path) and that will break the functionality for anyone using that flag.

I would submit that anyone using that flag is an advanced user and probably isn’t launching the console from the docs.