feat: skip build for APM Server if there are no code changes
docs/build-pr is the slowest job in apm-server CI pipeline and it is
impacting significantly developer velocity when merging PRs.
Update the buildkite pipeline to exit early if there are no
docs changes.
APM Server docs is actually in a separate repository (observability-docs)
and the few remaining files are barely touched (except for the changelog)